starting to make tests

+++ client/offlineActor.js
@@ -19,6 +19,10 @@ class OfflineActor extends Actor{
         this.collections = {}
     }
 
+    unsubscribe(ticket){
+        delete this.ticketFilters[ticket]
+    }
+
     subscribe(ticket, args){
         let predicate = args.shift()
         let collection = store.getCollection(predicate)
@@ -28,7 +32,7 @@ class OfflineActor extends Actor{
         }
 
         let filter = this.filters[predicate](...args)
-        this.ticketFilters[ticket] = filter
+        this.ticketFilters[ticket] = {predicate, filter}
         for(let key of Object.keys(this.collections[collection])){
             let doc = this.collections[collection][key]
             if(filter(doc)){
@@ -37,30 +41,35 @@ class OfflineActor extends Actor{
         }
     }
 
-    handle(key, value, collection){
+    handle(doc, collection){
         if(!this.collections[collection]){
             this.collections[collection] = {}
         }
-        let oldDoc = this.collections[collection][key] || null
-        let newDoc = Object.assign({}, oldDoc, value)
+        let oldDoc = this.collections[collection][doc.id] || null
+        let newDoc = Object.assign({}, oldDoc, doc)
         for(let t of Object.keys(this.ticketFilters)){
-            let filter = this.ticketFilters[t]
-            let type
-            if(!oldDoc || !filter(oldDoc)){
+            let {filter, predicate} = this.ticketFilters[t]
+            let type = null
+            if((!oldDoc || !filter(oldDoc)) && filter(newDoc)){
                 type = 'add'
-            }else if(!filter(newDoc)){
+            }else if(oldDoc && filter(oldDoc) && !filter(newDoc)){
                 type = 'delete'
-            }else{
+            }else if(oldDoc && filter(oldDoc) && filter(newDoc)){
                 type = 'update'
             }
-            this.ws.tell('dispatch', {ticket: parseInt(t), type, data: {newVal: newDoc}, predicate: 'todos'})
+            if(type) {
+                this.ws.tell('dispatch', {ticket: parseInt(t), type, data: {newVal: newDoc}, predicate})
+            }
         }
-        this.collections[collection][key] = newDoc
+        this.collections[collection][doc.id] = newDoc
     }
 
     send(obj){
         let collection = obj.args[0]
         switch(obj.type){
+            case 'unsubscribe':
+                this.unsubscribe(obj.ticket)
+                break
             case 'subscribe':
                 this.subscribe(obj.ticket, obj.args)
                 this.ws.tell('dispatch', {type: 'ready', ticket: obj.ticket})
@@ -70,13 +79,13 @@ class OfflineActor extends Actor{
                 let doc = obj.args[2]
                 doc.id = id
                 this.ws.tell('dispatch', {type: 'rpc', data: 1, ticket: obj.ticket})
-                this.handle(id, doc, collection)
+                this.handle(doc, collection)
                 break
             case 'add':
                 let ret = {type: obj.type, ticket: obj.ticket, data: obj.args[1]}
                 ret.data.id = ':' + obj.ticket
                 this.ws.tell('dispatch', {type: 'rpc', data: 1, ticket: ret.ticket})
-                this.handle(ret.data.id, ret.data, collection)
+                this.handle(ret.data, collection)
                 break
         }
     }
